Implement Maven version parser. (#248)
This is a scheme with an ambiguous and complicated spec. This was
implemented based on both this and Maven's implementation in Java to
fill in the gaps.

There doesn't appear to be an existing good Python library.

This will be hooked up to the list of supported ecosystems in a future
PR to enable Maven vulnerability ingestion.