Packagist version enumeration support (#675)
Fixes #230 

Potential improvements before merging:
- From the Best Practices here: https://packagist.org/apidoc, it is
recommended to send a User-agent header to allow them to contact us to
resolve any bandwidth issues.
- Also can utilize `If-Modified-Since` headers to avoid expiring caches
too early when calling the API
- The version enumeration is not quite identical to the C implementation
in PHP, though matches the slightly ambiguous specification given in the
documentation. Should be fine for the vast majority of cases.

We should also think about pushing the packagist comparison upstream to
the univers library.