OSV scanner github action (#698)
Creates docker image for osv scanner to allow it to run as a github
action, and the appropriate action.yml.

The current implementation is not particularly useful, only printing out
the found vulnerabilities to the logs. It should be running under this
PR.

TODO: Decide what the goal of the github action will be.