Revert git shallow clone (#772)
pygit2 does not support shallow clones at the moment, causing an error
when attempting to walk shallow clones.

Also adds some minor logging improvements to provide better visibility
of what is happening.

This does also revert a bunch of helpful typing and documentation
options. For now it might be better to leave them reverted, as it will
make it much easier to reapply the shallow clone changes once support is
added in pygit2.