feat: implement `--recursive` flag (#816)
As mentioned
[here](https://github.com/google/osv.dev/pull/815#issuecomment-1311245059),
I've made the default to be checking only the files in the target
directories, as I think that is the better UX: generally easier to
manage inclusion in userland (i.e. loops, globs, `tree`, etc) than
exclusion (case and point, right now there is no way to exclude anything
from the scanner).

I suspect it is probably going to be best to replace `filepath.Walk`
with a recursive loop, at least if `recursive` is not passed because
otherwise I believe you load the whole file tree into memory every time.

I have started writing some tests for this, but will do a seperate PR
for that since it requires more unrelated work.