Do not filter ecosystem-less vulnerabilities
OSV records converted from NVD CVE entries do not have a package,
therefore they do not have an ecosystem, and are getting filtered out.

This adds a test for this type of vulnerability, and modifies the
ecosystem filtering behaviour to accommodate these vulnerabilities.

I also discovered that the existing
test_update_partly_bad_ecosystem_delete  was not indented correctly and
not being run, and when it is, fails unexpectedly.