Use the precomputed CPE-to-repo map as a cache
Look up the repo mapping first, before doing CVE-specific reference
examination.

For the 2022 CVE data, and the current state of the CPE Dictionary,
there's still an additional 1,629 potential repositories picked up from
CVE references, and an additional 210 OSV records generated, so it's
premature to exclusively rely on the CPE Dictionary.