Don't run the integration tests on every PR
The integration tests can randomly fail because they're operating on
live data that can change unexpectedly. This results in tests failing on
a PR through no fault of the code in the PR. We already run the
integration tests on every *commit*, so it's just a matter of surfacing
those failures to the people who can do something about them.