Update deployment script and pipenv.lock (#968)
Updates pipenv.lock, for some reason the pipenv.lock was considered out
of date, maybe we need to restrict the versions on our pipenv files more
to avoid this in the future.

Also updates deploy script to not check origin/master, but the remote
`github.com/google/osv.dev` specifically, this is required because the
upstream remote might be named differently from `origin` (e.g. I name my
fork `origin`, and the main repo `upstream` instead).

And fix issue with date checks where bash assumes it's consuming hex
number instead of decimal number because of the leading 0.