Make the Maven tests use deps.dev
Avoid leaking the API key for deps.dev in test output.

This is an interim measure to address Maven flakiness causing
intermittent test failures and won't work when we move tests over to
GitHub Workflows. The ultimate solution is to make the tests hermetic
and mock out the network dependencies.