Use a Cloud Secret for the deps.dev key
Turns out kubectl isn't available in the ci Docker image so replicate
the deps.dev API key as a Cloud Secret and retrieve it that way.

Make the test execution more verbose so it's very clear which ones are
being skipped and add a failsafe for the Secret not being successfully
retrieved.