Automate Terraform for oss-vdb-test (#1017)
Adds a number of things needed to start automatically running terraform
changes on the staging environment.

- Adds Terraform Docker image, usable by Cloud Build to run Terraform
for our project.
- Adds a cloud build trigger on `oss-vdb` to run `terraform plan` on
`oss-vdb-test` when terraform files are changed
  - Output is only viewable in Cloud Build at the moment
- Adds a cloud build trigger on `oss-vdb` to run `terraform apply` on
`oss-vdb-test` on pushes to the master branch.
- The cloud build yaml will eventually be used for building Docker
images & running Cloud Deploy

The Cloud Builds use the existing `deployment` service account from
`oss-vdb`. IAM permissions have been granted to it on `oss-vdb-test` to
allow it to modify the project as needed.