Refactor repo denylisting for reuse
Interim move to allow the denylisting to be used more broadly by baking
it into the repository URL extraction and normalization function.

Added benefit, better testing.

Unexpected bonus of that: discovered repositories hosting their own
GHSA's were being incorrectly ruled out of scope, so that's a bit of a
win...