Update local ssd entry again (#1201)
Add local ssd entries to the disk, minimum is 4. Each disk is 375 gb for
1.5tb of local storage per node.

This should fix the out of disk space errors.

It seems like by default if there are local ssd's attached, then
emptyDir uses the local ssds. It's really hard to find solid
documentation on this though.