Docs conversion PR #2: Add redirects (#1204)
[This must be merged
first](https://github.com/google/osv.dev/pull/1203).

This is the second in a multi-part pull request to migrate OSV.dev
documentation.

This pull request adds redirects to the newly deployed docs. 

Old docs are still in the codebase, but they will be bypassed. 

In the final pull request, they will be deleted.

Co-authored-by: Oliver Chang <oliverchang@users.noreply.github.com>