Use deps.dev v3alpha HTTP API instead of v2alpha gRPC API
deps.dev recently released an updated API, which is available via HTTP
as well as gRPC, and does not require authentication.

Update DepsDevMixin to use the new HTTP API when enumerating versions,
and remove the fallback logic for Maven. Remove the old protos, along
with the logic for handling API keys.

Addresses #1211.