Change logging target for vulnfeeds to be dynamic (#1241)
Change logging target for vulnfeeds depending on what environment it's
running in.

This changes:
- alpine
- cve-downloader
- combine-to-osv

to get their project from the environment instead of hard coding. (Added
a extra utility function to create the Logger wrapper and retrieve the
project from the env).

This does not change cperepos to avoid merge conflicts.

Then changes almost all of the yaml configs to include the
GOOGLE_CLOUD_PROJECT environment variable (this includes the cperepos
yaml files, can undo that if it causes merge conflicts @andrewpollock )


Also fixes alpine converter not using the logger wrapper for invalid
versions.