Bucketed Indexer (#1236)
Indexer implementing bucketing to speed up and make more time consistent
queries.

Changes the API interface to add 2 extra fields in the response.

- minimum file matches
- estimated different files

Also added a number of bug fixes:

- Commit hash is now using the real commit hash rather than the
reference hash of the tag. The datastore key is still using the
reference hash for cases where multiple tags are on the same commit. An
extra field `reference` in the datastore to save the reference hash
- Controller now correctly avoids pushing already existing items to the
pub/sub
-