Add typing, add timeout, add kernel quick fail (#1290)
Added the quick fail for Kernel queries, and tried matching the ndb
query timeout with the GRPC time_remaining function. Added a log to
check if that function does what I think it does and matches what we set
in cloud run.

Also added a bunch of type hints just to make development more
efficient.