Fix indexer occasional error, change commit response to hex (#1346)
Make the indexer API resilient to cases where buckets are created but
RepoIndex entry is not (this might happen when something fails when
writing to the datastore), these errors should only be temporary and be
resolved on the next indexer run.

Added a warning message there so we can observe how often this occurs. 

Also changed commit in the response from base64 to hex.