Skip examining unrelated CPEs on CVEs
Problem: sometimes a CVE can have multiple unrelated CPEs on it, a
pathological example being CVE-2022-27778, which is for a curl
vulnerability, but also affects MySQL.

This changes the behaviour to effectively only focus on the first CPE in
the configurations of the CVE.

An alterative treatment of CVEs like CVE-2022-27778 would be to emit
multiple affected[].ranges[] entries.

This also refactors the VendorProduct struct into the cves package for
centralised use.