Remove unavailable package (#1375)
The link checker caught another dead link. The previously linked go
package was recently deprecated and is no longer available.

If we wanted to, we could replace this link with a link to the [Go vuln
database](https://go.dev/security/vuln/database) but I don't think that
is the intention of this section. I am inclined to just delete the link.