Change aliases for debian to related instead (#1381)
Debian advisories' CVE lines lists multiple CVE's that's fixed by the
new version. This is not what aliases field is used for, and is better
suited to the related field.

This will have the consequence of not being able to search by the CVE on
the osv.dev website to find the DSAs, since only aliases are currently
placed in the search_indices, not related. Created #1380 to track this.