Adjusted link checker to run as cron (#1386)
I have adjusted this so it runs once daily. 

It also runs on pushes, but only scans any files that have been
modified. Failures on push should not prevent merging, but we can remove
this in the case it does.

If the link checker finds something when you push, you can either
address it in your pull request, or you can open an issue and I will fix
it.