Add a log message for a task's latency
Add a new timestamp attribute to pubsub messages emitted by the importer

Then if they are present, have the worker log how much time passed
between when the messages were published and when the tasks were
completed.

This will allow for a log based metric to measure how long things are
taking, notably when it comes to importing, for SLO measurement.