Increase deadline for combine-to-osv
Including the NVD input seems to causing the job to run into the current
1h deadline and get terminated, double the deadline to get a sense of
how long it is taking to complete and what optimisations might be
necessary.

Annoyingly this wasn't being very well surfaced in the logs, and
additionally the concurrency policy seemed to be dueling with the
deadline, causing further confusing log entries.