Avoid an empty package field in the resulting OSV record
Despite the

```
Package  *AffectedPackage `json:"package,omitempty"`
```

I think that because it's a pointer to a struct with default values,
it's not being considered "empty" so it doesn't get omitted, and the
record winds up with an unsightly looking

```
package: {}
```

which seems to be further angering the importing of it.

Improve the existing tests for this, including the PURL one, which was
not testing correct behaviour.