Ease fuzzy tag matching by removing repo name (#1599)
There's a particularly hairy fuzzy tag matching scenario that doesn't
lend itself easily to the regex, where the repo name ends in a digit and
the tags have this in their prefix. It leads to incorrect normalization
of the tag. So try and identify such cases and preemptively remove the
name-based prefix from the tags to be normalized.