Store an empty ecosystem as a null in Datastore
Datastore does not support searching for an empty array, so for cases
where it is empty, set it to None, which will translate to a null.

This will allow for manual searching of the Bug Kind for ecosystem set
to null (finding only the converted NVD CVEs being the primary use case).