Refactor vulns.CVEIsDisputed to use a local clone and speed up execution (#1621)
This change makes consulting the CVE List to determine dispute status
(for marking converted CVEs as withdrawn) conditional and consult a
local Git clone rather than the CVE
List remotely, in an attempt to speed up the overall run time of
`combine-to-osv`.

It refactors `vulns.CVEIsDisputed` to consult a local clone of the CVE
List rather than make numerous individual HTTP requests.

This should significantly speed up and improve the reliability of
checking CVEs to see if they are disputed.