Add validation for name/version with purl in query (#1652)
Closes https://github.com/google/osv.dev/issues/1451

I've added some validations for when the query contains both a
`package.purl` and a `package.name`, and when the `package.purl`
contains a version and there is a redundant `version` included in the
query. In both cases, we're raising a 400 with an error message. Also
I've moved the `ecosystem` validation into `do_query` to keep all the
validations in one place.

---------

Co-authored-by: Andrew Pollock <andrewpollock@users.noreply.github.com>