Remove unnecessary symlinks to `osv` (#1716)
I'm pretty sure that these symlinks to the `osv/` folder in `api` &
`appengine` have been unnecessary. Both Pipfiles use the `osv` module at
the root of the repo, and I can't seem to find anything that traverses
into the symlinks.

I'm not entirely sure why we had them in the first place?

There's another one at `gcp/api/v1/osv`, but it seems to be used to when
we generate the protobufs.