Rework everything to use NVD API 2.0 format data
Added addtional instrumentation to assist with troubleshooting failures

Make API 2.0 downloading more fault tolerant:

Retry individual pages if a 503 is encountered or a potentially
malformed page is served.

Add jq to the Docker images as it is used for
slicing up the retrieved output by year.

Simply how test data is managed for ease of expansion.