Make API liveness probe more lenient (#1952)
Large batch queries can impact how long other queries running on the
same instance take, making the liveness queries take possibly over a
second to respond and failing the liveness checks.

I've bumped the probe `timeoutSeconds` up from the default 1 to 5 to see
if this reduces the number of failures that happen. I've also explicitly
defined the `failureThreshold` and `periodSeconds`, but I've set them to
the defaults values.