Handle GCS bucket record deletion
Detect and plumb through to worker for deletion from Datastore Bugs that
are no longer present in GCS.

Related changes (#2023, #2029) to the NVD CVE generation cause CVEs no
longer being generated (due to changes in the heuristics like that made
in #1707) to remain existing in GCS.

This PR addresses cases like this and the need identified in #1467 by
adding a deletion phase to the importing of new/updated records. The
functionality is flag-protected, it won't go live in Production until
a new `--delete` flag is included so in the execution of the importer.

Incidental changes:

- use the GCS bucket `directory_path` to efficiently filter the blobs
  returned when listing bucket contents
- make blob retrieval resilient to blob generation change between blob
  listing and blob retrieval (this can happen if `combine-to-osv`
  happens to have run in between these two points in time)
- fix a behavior inconsistency with schema validation not being
  performed when `ignore_last_import_time` is in effect (addressing
  head scratching TODO from @michaelkeder)
- tidy up the existing tests, making them more readable and debuggable
- add a slow manual test against live data in staging to validate
  real-world behavior and run time (this adds ~13 minutes to an import
  run on just the CVE GCS bucket)