Adjust how converted records are copied
The previous copying technique optimised for getting the records to GCS
as fast as possible for opportunistic ingestion by combine-to-osv.

The problem with this approach is that if any improvements to the
conversion code stop generating records that shouldn't be generated, the
most recent copy lingers indefinitely. This is causing situations like
the one described in #1961

So collect all of the records generated and copy them to GCS at the end
of the complete run, deleting anything in GCS that isn't still being
generated.