Sanity check to prevent duplicate commits (#2127)
This adds a check for the same commit being added to more than once when
converting versions to commits.

#1984 surfaced that OpenSSL CVEs are being converted with the same
commit as both `introduced` and `fixed` (due to incorrect normalization
of OpenSSL versions). While the underlying defect is being addressed,
add some sanity checking so that if the same commit hash can't be added
to a GIT range being generated when converting versions to commits.