Improve browser navigation back for vulnerability list page (#2144)
Issue: #1697 

As mentioned in the issue, the vulnerability list page gets stuck after
apply filters and click on the browser's back button, I've done some
research on this but can't find a nice solution, and I feel this could
be a weird bug of Turbo. So I just disabled the Turbo cache behavior for
vulnerability list page, and it just reloads the page after click the
back button.

Please let me know if you have a better suggestion.