V2 UI: Handle overflowing tabs on vulnerability details better (#2134)
Issue: #381 

Since [spicy-sections](https://github.com/tabvengers/spicy-sections)
doesn't have a feature of expanding all collapsed sections by default, I
decided to programmatically click on those package headers when the page
is loaded. But this method is bit of awkward, so let me know if you have
a better idea.

Mobile view:

![mobile-view](https://github.com/google/osv.dev/assets/13760813/1efe57ee-0c29-4f8e-a041-99de7675f6cd)

An alternative way of implementing similar feature is using
`exclusive-collapse` mode for `spicy-section`, but in that mode only the
first section is expanded by default, and users are only able to expand
one section (which is kind of similar with the tab-bar mode). Here is a
demo of different modes of Spicy Section, and you just need to click on
the "exclusive-collapse" button to check exclusive-collapse mode:
https://spicy-sections.glitch.me/programatic.html