Sort the AffectedCommits
Instability in this data structure adds unnecessary noise.

There is currently variability in the output, which is resulting in
unnecessary copying to GCS (because the checksum of the file changes),
and we're trying to get the true modification time of a resultant OSV
record to reflect that of the newest constituent input part to
`combine-to-osv`.