Add parallel exporter (#2167)
Exporter currently take too long to complete (around 1 hour and 10
minutes), and will only get longer to run as the OSV database expands.

This change parallelizes the exporter along each ecosystem, separating
the ecosystem export portion of the script from the selection of each
ecosystem. This roughly reduced the time taken down to how long the
longest ecosystem takes to export, 13 minutes or so.