Normalize GitHub repos to lowercase
Problem: #2182 made checks for existing commits case-insensitive by
repo, but left the underlying repo string alone. There are a lot of
mixed-case GitHub repos in existence, because cves.extractGitCommit()
takes the repo verbatim.

vulns.AddPkgInfo() aggregates events by repo, case insensitively, so was
producing incorrect GIT events.

GitHub repo names are known to be case insensitive, so this is safe for
them. It's definitively less safe for other URLs, so limit to just them
for now.