Fix for a small conversion regression
present in the VendorProductToRepoMap (intentionally).

This also means that CVEs for repos that lack tags but otherwise have a
commit hash in the CVE's references now get excluded from
consideration from conversion.

CVE-2023-0327 before #2233 had an entry in the VendorProductToRepoMap
for `theradsystem_project:theradsystem`,
even though the repo wasn't usable. Afterwards, it does not:

```
[CVE-2023-0327]: Failed to derive any repos for "theradsystem_project" "theradsystem"
[CVE-2023-0327]: Summary: [CPEs=1 AppCPEs=1 DerivedRepos=0]
[CVE-2023-0327]: Passing due to lack of viable repository
```

Add test coverage for `ReposFromReferences()` and fix a couple of bugs
that it flushed out for when there is no VendorProductToRepoMap cache in
use.

Remove some unnecessary code stubs that aren't going to get filled out
(the approach to determining CVE scope changed to trying them all,
regardless of language)