Fix another bug in alpine enumeration (#2241)
Turns out alpine enumeration currently works by accident, relying on git
to bundle the changes to package version and package revision in the
same diff hunk, which is not always the case.

This caused enumeration to break for some packages (e.g. busybox
https://osv.dev/vulnerability/CVE-2023-42366 , 3.16.1 only have r11).

This change fixes this by separating out the "start" search for versions
and revisions, so changes to both are displayed, not just changes to the
first element found (which is usually `pkgver`, therefore it was relying
on changes in `pkgrel` to be in the same diff hunk in git to be shown).