Update snapshots (#2256)
Turns out we haven't regenerated snapshots in a while, and PrettyPrinter
at some point changed behaviour to also add the indent to the start of
an array/list.

Also removed the custom max_width in prettier I added 8 months ago. I'm
not exactly sure what that was for, but it causes very large changes to
the existing snapshots, while also not being significantly prettier.
This is what causes the big change to the REST snapshots in this PR, as
they were generated when the width=200 was set.