Expand examples of precision
Feedback was that it was possible to perceive this as known unfixed
vulnerabilities as being of lower quality than ones with known/available
fixes, and that is not the case, as such known unfixed vulnerabilities
can and do legitimately exist (as is the case when the vulnerable
package is orphaned). These are a legitimate true positive with no
available fixed version.