Fix a bug that causes premature termination of processing (#2283)
It turns out that trying to reimport a subsequently non-existent record
(i.e. when the originating CVE was rejected), the source file no longer
exists in GCS and the retrieval failure raises an exception, which is
being silently caught, which then caused the transaction to terminate.

So don't silently catch exceptions, other than the general one used to
terminate in dry-run mode.

Also removed unnecessary copy-pasta for this program's use case (it's
hard-limited to 30 records at a time due to Datastore query limits, so
batching makes no sense).