Display the correct introduced tooltip
Prior to this change, when the introduced version
in the affected range was 0, the tooltip displayed:
"The exact introduced commit is unknown".

This change, updates the html block to be more
readable, and displays a more descriptive tooltip
based on the type of the range, indicating if it has
versions or commits.