Replace relative time tag with template filters
Currently, we use relative-time package to convert
timestamp to a human readable relative time, when
displaying published time of the bug in the list page.

Using this approach is not optimized as it changes
the texts in the loading time and leads to CLS issues.

This change replaces relative-time lib in favor of a
new template tag which converts the published to a
human readable relative time using a utility function.