Encourage use of scorecard.dev by third-party integrators
We'd like to ensure we're linking to high quality security
projects/products and believe that the OpenSSF's Scorecard helps
encourage that.

Also promote the OpenSSF's Concise Guide for Evaluating Open Source
Software to consumers of the third-party software we link to.

Small steps to avoiding unintentionally linking to potentially malicious
or poor-quality security software.