Add durable support for overriding the bucket
This is a more durable solution to the edge case we were trying to
correct from, where the Bug in Datastore was from a GIT source (with a
different path to the source file) than what we wanted to trigger
reimporting for, and so various sanity checks and assumptions fail.

This adds support for overriding the bucket name (and optionally, the
path within the bucket), which is normally self-determined from the
existing record.

```
pipenv run python tools/datafix/reimport_gcs_record.py --verbose --no-dry-run --project oss-vdb --bucket cve-osv-conversion:osv-output CVE-2022-2816
```