refactor: Migrate from Pipenv to poetry (#2460)
This moves Pipenv to poetry, and removes Pipfile Pipfile.lock, and other
mentions or Pipenv from the repository.

Changes in addition to a direct migration:
- Appengine
- Removes gcloud appengine related files, as we are now fully on cloud
run via building a docker image (With the exception of index.yaml, which
is for appengine datastore)
- Add a `build.sh` script for building the docker image similar to other
`build.sh` scripts.
- Removed most dependencies in deployment as it's originally only needed
for appengine (see the deleted `gcp/appengine/deploy.sh`)
- Python environment
- For many of the docker images, pipenv installs to the **system**
python dependencies. This is not really ideal, as even in a docker
container, it is good practice to compartmentalise into separate
virtualenvs.
- For the worker image, I have retained this behaviour, and will leave
this migration to a future PR, as it affects a large number of different
images based on it, e.g.:
    - ci
    - cron
    - exporter
    - importer
- For other non worker based images, I have switched to using
virtualenvs, this includes:
    - Appengine/Cloud Run
    - API
    - Debian (already uses virtualenvs)


TODO/looking for suggestions:

The poetry version is defined in every file that it is installed with
`==`, is there a way to set the version number for the whole repository
in one place?