fix: unpin github pages dependency (#2489)
Unpin the GitHub pages dependency, as GitHub pages always use the latest
version of this dependency when publishing pages on GitHub, so this
Gemfile is only for local development. By having this as always the
latest version, we can see an accurate version of the published page
when building locally.