fix(importer): address non-deterministic BucketImporterTest.test_bucket() failure
In the end, the problem was:
- `assert_has_calls(calls, any_order=False)` being the source of the problem

This assertion doesn't mind if there's *extra* calls above and beyond
what is being checked for, but without `any_order=True`, it *will* fail
if the results are out of order.

The *ordering* was the problem, with CVE-2022-0128 always being in the
set of calls, but not always interfering with the order of the other calls
being actively checked for. The ordering of the calls is unavoidably
non-deterministic (and fine) because they're being made from separate
threads.