chore(frontend): improve frontend vuln debuggability
I hit weirdness with an incomplete vulnerability page, which took longer
than I'd have liked to get to the bottom of. It was because the source
`redhat-stage` was renamed to `redhat` AND the particular record I was
looking at had failed to import successfully from the newly named
`redhat` source.

This scenario is unexpected enough that it shouldn't be something that
happens silently.