fix(worker): handle Datastore exceptions (#2754)
Pathologically large records can be too big for Datastore, resulting in
"entity is too big" errors and unhandled exceptions. This is happening
for a few of the Red Hat records at present.

This commit handles failures to `put()` the bug in Datastore, and logs
the exception with context as to the offending record, to ease
debugging.