feat: make critical batch jobs fail more noisily
This commit adds an error handler to the wrapper script that runs the
batch job workload, so that if it fails unexpectedly due to being run
with `set -e`/`set -u` we stand a better chance of GCP's error reporting
alerting us to it, or at least having additional context in the logs.

This is because until we get proper k8s job monitoring implemented,
we're solely dependent on error reporting alerting us to job failures.