feat(importer): prepare for selective strict mode enablement (staging)
This commit explicitly sets strict mode on a per-repository basis. This
change is a no-op until the importer's `--strict_validation` flag is
also set to True, and it will then be a no-op for all sources other than
the `cve-osv` one where we can trial having it enabled.

Part of #2188