feat(frontend): improved UX for missing vulnerabilities
This commit modifies the 404 flow so that:
- when a request for a non-existent vulnerability is received, the 404
  page reflects the nature of the failed request
- when it is known that the vulnerability failed to import, this is
  explicitly surfaced

A new FAQ entry discussing this is also added.

Additionally, the 404 generation was tidied up to make it easier to
understand where the 404 comes from. There were many unreachable paths
that called `abort()`, which made the code unnecessarily unreadable.

It also addresses requests for `/favicon.ico` generating 404's from the
redirector by adding an explicit handler for it.

Part of #2190